+++ crates/agent/src/tools/edit_session/reindent.rs
