+++ crates/agent/src/tools/edit_session/streaming_fuzzy_matcher.rs
