+++ crates/agent/src/tools/edit_session/streaming_parser.rs
@@ -1,6 +1,6 @@
 use smallvec::SmallVec;
 
-use crate::tools::edit_session::{Edit, PartialEdit};
+use super::{Edit, PartialEdit};
 
 /// Events emitted by `StreamingParser` for edit-mode input.
 #[derive(Debug, PartialEq, Eq)]
